IVYPORTAL-13452: Support call original case information when using iframe - Handled feedback: changed type to Long

+++ AxonIvyPortal/PortalTemplate/src/ch/ivy/addon/portal/generic/bean/IFrameTaskTemplateBean.java
@@ -63,7 +63,7 @@ public class IFrameTaskTemplateBean extends AbstractTaskTemplateBean implements 
   private boolean isWorkingOnATask = true;
   private Map<String, Object> overridePortalGrowlMap = new HashMap<>();
 
-  private Integer caseId = null;
+  private Long caseId = null;
   
   private PortalNavigator navigator = new PortalNavigator();
 
@@ -159,7 +159,7 @@ public class IFrameTaskTemplateBean extends AbstractTaskTemplateBean implements 
     isHideCaseInfo = Optional.ofNullable(requestParamMap.get(IS_HIDE_CASE_INFO)).map(BooleanUtils::toBoolean).orElse(false);
     isHideTaskName = Optional.ofNullable(requestParamMap.get(IS_HIDE_TASK_NAME)).map(BooleanUtils::toBoolean).orElse(false);
     isHideTaskAction = Optional.ofNullable(requestParamMap.get(IS_HIDE_TASK_ACTION)).map(BooleanUtils::toBoolean).orElse(false);
-    caseId = Optional.ofNullable(requestParamMap.get(CASE_ID_PARAM)).map(p -> StringUtils.isNotBlank(p) ? Integer.parseInt(p) : null).orElse(null);
+    caseId = Optional.ofNullable(requestParamMap.get(CASE_ID_PARAM)).map(p -> StringUtils.isNotBlank(p) ? Long.parseLong(p) : null).orElse(null);
     isWorkingOnATask = Optional.ofNullable(requestParamMap.get(IS_WORKING_ON_A_TASK)).map(p -> StringUtils.isNotBlank(p) ? BooleanUtils.toBoolean(p) : true).get();
   }
 
@@ -213,11 +213,11 @@ public class IFrameTaskTemplateBean extends AbstractTaskTemplateBean implements 
     return isWorkingOnATask;
   }
 
-  public Integer getCaseId() {
+  public Long getCaseId() {
     return caseId;
   }
 
-  public void setCaseId(Integer caseId) {
+  public void setCaseId(Long caseId) {
     this.caseId = caseId;
   }
 }
